+++ config/me_cms.php
@@ -131,10 +131,6 @@ return ['MeCms' => [
 	],
 	//Security
 	'security' => [
-		//Array of banned IP addresses.
-		//You can use the asterisk (*) as a wildcard.
-		//With "false" or an empty value, access is granted to any ip addresses (no limitation).
-		'banned_ip' => [],
 		//Link for "IP map". The `{IP}` string will be replaced with the IP address
 		'ip_map' => 'http://www.traceip.net/?query={IP}',
 		//Link for "IP who is". The `{IP}` string will be replaced with the IP address